refactor(buttondown): move newsletter API boundary
